third_party/WebKit/LayoutTests/platform/android/virtual/trustedeventsdefaultaction/fast/events/keyboardevent-key-expected.txt

Tests KeyboardEvent.key attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Modifier Keys
PASS lastKeyboardUpEvent.key is 'Control'
PASS lastKeyboardDownEvent.key is 'Control'
PASS lastKeyboardDownEvent.key is 'Shift'
PASS lastKeyboardDownEvent.key is 'Shift'
PASS lastKeyboardDownEvent.key is 'Alt'
PASS lastKeyboardDownEvent.key is 'AltGraph'
PASS lastKeyboardDownEvent.key is 'NumLock'
PASS lastKeyboardUpEvent.key is 'NumLock'
Whitespace Keys
PASS lastKeyboardDownEvent.key is 'Enter'
PASS lastKeyboardDownEvent.key is 'Tab'
PASS lastKeyboardUpEvent.key is 'Tab'
Navigation keys
PASS lastKeyboardDownEvent.key is 'ArrowRight'
PASS lastKeyboardUpEvent.key is 'ArrowRight'
PASS lastKeyboardDownEvent.key is 'ArrowDown'
PASS lastKeyboardDownEvent.key is 'ArrowLeft'
PASS lastKeyboardDownEvent.key is 'ArrowUp'
PASS lastKeyboardUpEvent.key is 'ArrowUp'
PASS lastKeyboardDownEvent.key is 'Home'
PASS lastKeyboardDownEvent.key is 'End'
PASS lastKeyboardDownEvent.key is 'PageUp'
PASS lastKeyboardDownEvent.key is 'PageDown'
PASS lastKeyboardUpEvent.key is 'PageDown'
Editing Keys
PASS lastKeyboardDownEvent.code is 'Backspace'
PASS lastKeyboardDownEvent.key is 'Insert'
PASS lastKeyboardUpEvent.key is 'Insert'
PASS lastKeyboardDownEvent.key is 'Delete'
Device Keys
PASS lastKeyboardDownEvent.key is 'PrintScreen'
PASS lastKeyboardUpEvent.key is 'PrintScreen'
General Purpose Function Section
PASS lastKeyboardDownEvent.key is 'F1'
PASS lastKeyboardDownEvent.key is 'F6'
PASS lastKeyboardDownEvent.key is 'F12'
UI Keys
PASS lastKeyboardDownEvent.key is 'Escape'
PASS lastKeyboardUpEvent.key is 'Escape'
PASS lastKeyboardDownEvent.key is 'MediaApps'
PASS successfullyParsed is true

TEST COMPLETE
- 
